DS-3484: Provide the option to only run the tests required to generate the REST API documentation
