+++ dspace-spring-rest/pom.xml
@@ -144,6 +144,7 @@
             </build>
 
         </profile>
+
     </profiles>
 
     <build>
