+++ dspace-spring-rest/src/test/java/org/dspace/documentation/bitstream/BitstreamControllerDocumentation.java
@@ -39,7 +39,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 /**
  * Documentation test for the {@link org.dspace.app.rest.repository.BitstreamRestRepository}
  */
-public class BitstreamControllerDocumentationTest extends AbstractControllerIntegrationTest {
+public class BitstreamControllerDocumentation extends AbstractControllerIntegrationTest {
 
     protected String getRestCategory() {
         return RestModel.CORE;
